Add files via upload
This repository contains the code for a fully functional prototype of the Reddula website. The prototype includes basic frontend and backend functionalities, developed using HTML, CSS, JavaScript, PHP/ MySql. While this prototype showcases the core features and design of the website, we aim to further develop and enhance it in the near future, adding more advanced features, improving user experience, and integrating additional functionalities.